API/requirements.txt
fastapi==0.128.0
uvicorn==0.40.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=3.1.0,<5
jinja2>=3.1.0
